Add a debugging option
We have added a debugging option which turns on the LED when the reward trigger (sent by L&N code) is received by the Bpod. This allows to visualize when the Bpod is "allowed" to deliver reward.
To do: link the debugging option to a tick box in the parameter GUI